# openapi.json
{
  "components": {
    "schemas": {
...
      "ChoiceRequest": {
        "properties": {
          "option_id": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Option Id",
            "type": "string"
          }
        },
        "required": [
          "option_id"
        ],
        "title": "ChoiceRequest",
        "type": "object"
      },
      "ContextAnswer": {
        "properties": {
          "answer": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Answer",
            "type": "string"
          },
          "question": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question",
          "answer"
        ],
        "title": "ContextAnswer",
        "type": "object"
      },
      "ContextIntakeRequest": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/ContextAnswer"
            },
            "maxItems": 12,
            "title": "Answers",
            "type": "array"
          },
          "initial_context": {
            "default": "",
            "maxLength": 12000,
            "title": "Initial Context",
            "type": "string"
          },
          "role_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Role Name",
            "type": "string"
          }
        },
        "required": [
          "role_name"
        ],
        "title": "ContextIntakeRequest",
        "type": "object"
      },
      "ContextIntakeResult": {
        "properties": {
          "missing": {
            "items": {
              "type": "string"
            },
            "maxItems": 8,
            "title": "Missing",
            "type": "array"
          },
          "questions": {
            "items": {
              "type": "string"
            },
            "maxItems": 4,
            "title": "Questions",
            "type": "array"
          },
          "status": {
            "enum": [
              "needs_more",
              "ready"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "default": "",
            "maxLength": 8000,
            "title": "Summary",
            "type": "string"
          },
          "urgent_warning": {
            "default": "",
            "maxLength": 1000,
            "title": "Urgent Warning",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "ContextIntakeResult",
        "type": "object"
      },
...
      "DecisionAssessment": {
        "properties": {
          "better_alternative": {
            "default": "",
            "title": "Better Alternative",
            "type": "string"
          },
          "choice": {
            "minLength": 1,
            "title": "Choice",
            "type": "string"
          },
          "commentary": {
            "minLength": 1,
            "title": "Commentary",
            "type": "string"
          },
          "turn_index": {
            "title": "Turn Index",
            "type": "integer"
          }
        },
        "required": [
          "turn_index",
          "choice",
          "commentary"
        ],
        "title": "DecisionAssessment",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlayerView": {
        "description": "The only turn payload play endpoints may return \u2014 gm_state never leaks here.",
        "properties": {
          "epilogue": {
            "default": "",
            "title": "Epilogue",
            "type": "string"
          },
          "narrative": {
            "title": "Narrative",
            "type": "string"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/Option"
            },
            "title": "Options",
            "type": "array"
          },
          "visible_state_summary": {
            "default": "",
            "title": "Visible State Summary",
            "type": "string"
          }
        },
        "required": [
          "narrative"
        ],
        "title": "PlayerView",
        "type": "object"
      },
      "PlaythroughAnalysis": {
        "description": "LLM output schema for the post-game analysis of the player's choices.",
        "properties": {
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/DecisionAssessment"
            },
            "title": "Decisions",
            "type": "array"
          },
          "improvements": {
            "items": {
              "type": "string"
            },
            "title": "Improvements",
            "type": "array"
          },
          "outcome": {
            "minLength": 1,
            "title": "Outcome",
            "type": "string"
          },
          "overall": {
            "minLength": 1,
            "title": "Overall",
            "type": "string"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "title": "Strengths",
            "type": "array"
          }
        },
        "required": [
          "outcome",
          "overall"
        ],
        "title": "PlaythroughAnalysis",
        "type": "object"
      },
      "PlaythroughDetail": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role_name": {
            "title": "Role Name",
            "type": "string"
          },
          "scenario_id": {
            "format": "uuid",
            "title": "Scenario Id",
            "type": "string"
          },
          "scenario_title": {
            "title": "Scenario Title",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/TurnOut"
            },
            "title": "Turns",
            "type": "array"
          }
        },
        "required": [
          "id",
          "scenario_id",
          "scenario_title",
          "role_name",
          "status",
          "turns"
        ],
        "title": "PlaythroughDetail",
        "type": "object"
      },
...
      "PlaythroughOut": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role_name": {
            "title": "Role Name",
            "type": "string"
          },
          "scenario_id": {
            "format": "uuid",
            "title": "Scenario Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "turn_count": {
            "default": 0,
            "title": "Turn Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "scenario_id",
          "role_name",
          "status",
          "created_at"
        ],
        "title": "PlaythroughOut",
        "type": "object"
      },
      "PlaythroughReview": {
        "properties": {
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlaythroughAnalysis"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role_name": {
            "title": "Role Name",
            "type": "string"
          },
          "scenario_id": {
            "format": "uuid",
            "title": "Scenario Id",
            "type": "string"
          },
          "scenario_title": {
            "title": "Scenario Title",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/ReviewTurn"
            },
            "title": "Turns",
            "type": "array"
          }
        },
        "required": [
          "id",
          "scenario_id",
          "scenario_title",
          "role_name",
          "status",
          "turns"
        ],
        "title": "PlaythroughReview",
        "type": "object"
      },
...
      "ReviewTurn": {
        "properties": {
          "chosen_option_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chosen Option Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "gm_state": {
            "$ref": "#/components/schemas/GMState"
          },
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "is_final": {
            "default": false,
            "title": "Is Final",
            "type": "boolean"
          },
          "player_view": {
            "$ref": "#/components/schemas/PlayerView"
          }
        },
        "required": [
          "index",
          "player_view",
          "gm_state",
          "created_at"
        ],
        "title": "ReviewTurn",
        "type": "object"
      },
...
      "StartPlaythroughRequest": {
        "properties": {
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlayerContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "context_summary": {
            "default": "",
            "maxLength": 8000,
            "title": "Context Summary",
            "type": "string"
          },
          "role_name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Role Name",
            "type": "string"
          }
        },
        "required": [
          "role_name"
        ],
        "title": "StartPlaythroughRequest",
        "type": "object"
      },
      "SuggestActionRequest": {
        "properties": {
          "text": {
            "maxLength": 300,
            "minLength": 3,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "SuggestActionRequest",
        "type": "object"
      },
      "SuggestActionResult": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "turn": {
            "$ref": "#/components/schemas/TurnOut"
          }
        },
        "required": [
          "accepted",
          "turn"
        ],
        "title": "SuggestActionResult",
        "type": "object"
      },
      "TurnOut": {
        "properties": {
          "chosen_option_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chosen Option Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "is_final": {
            "default": false,
            "title": "Is Final",
            "type": "boolean"
          },
          "player_view": {
            "$ref": "#/components/schemas/PlayerView"
          }
        },
        "required": [
          "index",
          "player_view",
          "created_at"
        ],
        "title": "TurnOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "CYOA Scenario Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/playthroughs/{playthrough_id}": {
      "get": {
        "operationId": "get_playthrough",
        "parameters": [
          {
            "in": "path",
            "name": "playthrough_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Playthrough Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaythroughDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Playthrough",
        "tags": [
          "playthroughs"
        ]
      }
    },
    "/api/playthroughs/{playthrough_id}/abandon": {
      "post": {
        "operationId": "abandon",
        "parameters": [
          {
            "in": "path",
            "name": "playthrough_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Playthrough Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaythroughOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Abandon",
        "tags": [
          "playthroughs"
        ]
      }
    },
    "/api/playthroughs/{playthrough_id}/analysis": {
      "post": {
        "operationId": "analyze",
        "parameters": [
          {
            "in": "path",
            "name": "playthrough_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Playthrough Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaythroughAnalysis"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze",
        "tags": [
          "playthroughs"
        ]
      }
    },
    "/api/playthroughs/{playthrough_id}/choice": {
      "post": {
        "operationId": "choose",
        "parameters": [
          {
            "in": "path",
            "name": "playthrough_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Playthrough Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChoiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Choose",
        "tags": [
          "playthroughs"
        ]
      }
    },
    "/api/playthroughs/{playthrough_id}/regenerate": {
      "post": {
        "operationId": "regenerate",
        "parameters": [
          {
            "in": "path",
            "name": "playthrough_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Playthrough Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate",
        "tags": [
          "playthroughs"
        ]
      }
    },
    "/api/playthroughs/{playthrough_id}/review": {
      "get": {
        "operationId": "review",
        "parameters": [
          {
            "in": "path",
            "name": "playthrough_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Playthrough Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaythroughReview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review",
        "tags": [
          "playthroughs"
        ]
      }
    },
    "/api/playthroughs/{playthrough_id}/suggest-action": {
      "post": {
        "operationId": "suggest_action",
        "parameters": [
          {
            "in": "path",
            "name": "playthrough_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Playthrough Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestActionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suggest Action",
        "tags": [
          "playthroughs"
        ]
      }
    },
...
    "/api/scenarios/{scenario_id}/context-intake": {
      "post": {
        "operationId": "assess_context",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextIntakeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextIntakeResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assess Context",
        "tags": [
          "playthroughs"
        ]
      }
    },
...
    "/api/scenarios/{scenario_id}/playthroughs": {
      "get": {
        "operationId": "list_playthroughs",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlaythroughOut"
                  },
                  "title": "Response List Playthroughs",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Playthroughs",
        "tags": [
          "playthroughs"
        ]
      },
      "post": {
        "operationId": "start_playthrough",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartPlaythroughRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaythroughDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Playthrough",
        "tags": [
          "playthroughs"
        ]
      }
    },
...
  }
}